[fix] move canvas method and value
